requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
gpt4all==1.0.8
numpy==1.24.3
pillow==10.1.0
opencv-python==4.8.1.78
python-dotenv==1.0.0 